Extract examples to separate repo (websdk/nap-examples)
By removing the examples and putting them in their own repo, it's
easier to test against nap as though a regular user might. If we
keep the examples as part of this repository, we run the risk of
relying on internals, and so might not catch issues with published
versions.